+++ modules/api/api.py
@@ -99,6 +99,7 @@ class Api:
         self.add_api_route("/sdapi/v1/refresh-vae", endpoints.post_refresh_vae, methods=["POST"])
         self.add_api_route("/sdapi/v1/latents", endpoints.get_latent_history, methods=["GET"], response_model=List[str])
         self.add_api_route("/sdapi/v1/latents", endpoints.post_latent_history, methods=["POST"], response_model=int)
+        self.add_api_route("/sdapi/v1/modules", endpoints.get_modules, methods=["GET"])
 
         # lora api
         from modules.api import loras